backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
openai>=1.30.0
chromadb>=0.5.0
fastembed>=0.3.1
python-dotenv>=1.0.0
networkx>=3.3
numpy>=1.26.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
pydantic>=2.7.0
httpx>=0.27.0
